refactor: 사이드바, memo컴포넌트 newmemo하위로 이동
