--- apps/client/src/pages/home/home-page.tsx
@@ -1,5 +1,0 @@
-const HomePage = () => {
-  return <div>home-page</div>;
-};
-
-export default HomePage;
