--- apps/client/src/pages/home/index.ts
@@ -1,1 +1,0 @@
-export { default as HomePage } from './home-page';
