+++ apps/client/src/pages/new-memo/components/memo/ui/memo-input/memo-input.css.ts
