+++ apps/client/src/pages/new-memo/components/memo/ui/memo-input/use-navigation-blocker.ts
