+++ apps/client/src/shared/components/sidebar/sidebar.tsx
