--- apps/client/src/widgets/header/index.ts
@@ -1,1 +1,0 @@
-export { default as Header } from './ui/header';
